requirements.txt
# Django Core
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1

# GeoDjango
# GDAL version will match system GDAL
psycopg2-binary==2.9.9
djangorestframework-gis==1.0

# Authentication
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.27.1

# Utilities
python-decouple==3.8
python-dotenv==1.0.1

# Development
ipython==8.22.2
django-extensions==3.2.3
